Update Go dashboard
CPU and RSS memory panels are added at the top. The rate of heap
allocations is also added.

The 'pod_name' label is added to all panels.

The 'go_gc_duration_seconds' panel is removed because it is based on a
summary metric that can't be aggregated. Instead the rate of GC events
is added.

The 'go_threads' panel is removed, tracking the number of goroutines is
usually enough.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>