Update base images when building Kiali containers
The UBI7 base images used to build Kiali are, apparently, updated to patch for vulnerabilities, but the machine building the Kiali containers seems to have a "shared" docker daemon which is caching the base images. So, Kiali is being built with the cached base image retrieved at some point of the time.

To fix, adding --pull argument to check and get available updates of the base images at build time.

Although operator base images will also be checked for updates, the base image has a fixed version tag. So, updates are not retrieved. The :latest tag of the base image is the one with the security fixes, but I'm not sure if changing the base image would break the operator. Leaving that for another PR, unless I get the green light to change the base image in this PR.

Fixes #1781.