Move to Go1.12.13 (#1992)
* Move to Go1.13.4

With these changes, Makefiles will ask for a minimum go version of 1.13.4. Additionally, the Makefile is changed so that it's possible to specify the go binary using an environment variable; ex: `GO=go1.13.4 make build`. This is useful if user has installed several go version (see https://golang.org/doc/install#extra_versions).

This also is:
* removing references to istiooc, because it's no longer supported in Maistra;
* a quick README.md cleanup;

Fixes #1970.

* Move to use golangci-lint version 1.21.0

Because previous version 1.17.1 seems to not work in Go1.13. In consequence, this is also fixing the new linting errors.

* Move to use swagger version 0.21.0

Because previous version 0.19.0 seems to not work in Go1.13.

* Fallback to using Go1.12.13

go-swagger version >= 0.20.0 seems to be buggy and it's not working correctly with Kiali. Stepping back to using go1.12.13 which is compatible with go1.12.13.

* Fallback to using Go1.12.13

go-swagger version >= 0.20.0 seems to be buggy and it's not working correctly with Kiali. Stepping back to using go1.12.13 which is compatible with go1.12.13.

* Fallback to using Go1.12.13

go-swagger version >= 0.20.0 seems to be buggy and it's not working correctly with Kiali. Stepping back to using go1.12.13 which is compatible with go1.12.13.