Add an OpenID molecule test. (#50)
Fixes: https://github.com/kiali/kiali/issues/2882
This test requires a Minikube cluster to be running
and configured with an OpenID user of "admin@example.com" and password of "password"
and that user has permissions to see at least one namespace.
It requires an issuer URI at the minikube IP on .nip.io domain on port 32000. e.g. https://192-168-99-100.nip.io:320000
This works if you run "hack/k8s-minikube.sh --dex-enabled=true start"
and you run something like this to grant access:
  kubectl create clusterrolebinding openid-test --clusterrole=cluster-admin --user="admin@example.com"