Use new OpenShift interface for service serving certificates (#56)
* Use new OpenShift interface for service serving certificates

OpenShift is now requiring new annotations for the service serving certificates interface. This is doing the needed changes to use the new interface.

Part of kiali/kiali#2917.

* Use OpenShift service CA by default, if not specified in the CR

* Use new configmap for the CA bundle

OpenShift replaces the configmap when annotated with the special `service.beta.openshift.io/inject-cabundle=true` anntation. So an existent configmap cannot be re-used.

* Fix OS annotation and add delete kiali-cabudle CM on operand deletion

* Annotation value must be text. Forcing type by quoting.
* Add line to make sure that the kiali-cabundle config map is deleted when the Kiali CR is deleted.

* Remove kiali-cabundle config map only if OpenShift