Add feature flag for clustering/federation features
For now, this flag is only allowing or preventing creation of a role and a rolebinding that grants privileges for reading some secrets in the Istio namespace. There is no effect on Kiali server side, because it is failing gracefully. In case it's needed later, this flag is being passed to Kiali in its ConfigMap.

Default value is `true` for backwards compatibility.

Related kiali/kiali#4147.