For the upstream releases, we only will support "default" version (though technically v1.36 will work, too). If we want to support "v1.12" and "v1.24" then we will need to add this to the requirements.yaml file:
- name: community.kubernetes
  version: 1.2.1

I am not adding that now. We can add it later if, for whatever reason,
we need to support those really old versions.

The reason for this is because we lost the ability to use the
ansible operator base image that had this collection already
(we can't use that image because it doesn't support multiarch).