[operator] upgrade base image (#460)
* update operator to new base image and new ansible k8s collection
fixes: https://github.com/kiali/kiali/issues/4540

Update the base image to include the 1.10.1 (which is the multiarch upstream version of the 4.9 image).
Now also install the kubernetes.core v2 collection.

* switch default and v1.36 playbooks to use kubernetes/core v2

* kubernetes.core query behaves differently now - use lookup instead

* For the upstream releases, we only will support "default" version
(though technically v1.36 will work, too).
If we want to support "v1.12" and "v1.24" then we will need to add this
to the requirements.yaml file:

- name: community.kubernetes
  version: 1.2.1

I am not adding that now. We can add it later if, for whatever reason,
we need to support those really old versions.

The reason for this is because we lost the ability to use the
ansible operator base image that had this collection already
(we can't use that image because it doesn't support multiarch).