Fix auth molecules (#483)
* Fix auth molecules

Auth molecule tests are failing because of the new restrictions in the length of the encryption key. This is also causing the authorization code flow to be triggered, while molecules are testing the implicit flow. So, some changes in config are required for the molecules to pass correctly, which are conflictive between header and openid molecules. This fixes the conflictive config by changning the dex config for the openid-test molecule. The "default" dex config is the one that works OK for the header-auth-test molecule.

Fixes kiali/kiali#4682