Get custom_secrets to work with identity settings.
fixes: https://github.com/kiali/kiali/issues/4761

When creating a custom secret for Kiali identity, make sure this doesn't conflict with other secrets the operator creates.
This also removes the secret mount definition in the k8s template because on non-OpenShift clusters there is no default secret (the user must define their own via custom_secrets).
For OpenShift, only if the identity cert file is set to the default /kiali-cert/tls.crt will the default secret be mounted. This allows one to provide their own custom secret/cert on OpenShift - they just have to mount it in a directory not called /kiali-cert.