On minikube, I saw replicas not get reset to 1 after HPA was disabled. (#800)
I believe (though not 100% sure) it was due to a race condition caused by
the operator not deleting the HPA prior to updating the deployment.
This change now has the HPA being removed first.

This also fixes a flaky test seen on OpenShift when HPA is used.
Testing on CRC/OpenShift, there was a lag when HPA adjusts replicas
on the Deployment. This add a check in the molecule test so it waits
for the replicas to reach steady state before continuing.