bump community.general ansible collection to 9.0.0
this is necessary because Snyx is detecting a security error ("hardcoded secret") in the rax library.
Kiali doesn't use the rax library at all (it is a Rackspace related library), but just to
avoid any issues in the future with CVE / security scans flagging this again, let's just upgrade now.